Refactor: class has own file
